move sidebar components into dedicated pipeline directories
