+++ features/earn/guni/open/containers/GuniOpenMultiplyVaultForm.tsx
@@ -4,13 +4,13 @@ import {
   OpenMultiplyVaultStatus,
 } from 'components/vault/commonMultiply/OpenMultiplyVaultConfirmation'
 import { OpenMultiplyVaultTitle } from 'components/vault/commonMultiply/OpenMultiplyVaultTitle'
-import { SidebarOpenGuniVault } from 'components/vault/sidebar/SidebarOpenGuniVault'
 import { VaultAllowance, VaultAllowanceStatus } from 'components/vault/VaultAllowance'
 import { VaultChangesWithADelayCard } from 'components/vault/VaultChangesWithADelayCard'
 import { VaultErrors } from 'components/vault/VaultErrors'
 import { VaultFormContainer } from 'components/vault/VaultFormContainer'
 import { VaultProxyContentBox, VaultProxyStatusCard } from 'components/vault/VaultProxy'
 import { VaultWarnings } from 'components/vault/VaultWarnings'
+import { SidebarOpenGuniVault } from 'features/earn/guni/open/sidebars/SidebarOpenGuniVault'
 import { extractGasDataFromState } from 'helpers/extractGasDataFromState'
 import { useFeatureToggle } from 'helpers/useFeatureToggle'
 import { Trans, useTranslation } from 'next-i18next'
