+++ features/multiply/manage/sidebars/SidebarManageMultiplyVaultTransitionStage.tsx
