+++ features/multiply/open/containers/OpenMultiplyVaultForm.tsx
@@ -4,7 +4,6 @@ import {
   OpenMultiplyVaultStatus,
 } from 'components/vault/commonMultiply/OpenMultiplyVaultConfirmation'
 import { OpenMultiplyVaultTitle } from 'components/vault/commonMultiply/OpenMultiplyVaultTitle'
-import { SidebarOpenMultiplyVault } from 'components/vault/sidebar/SidebarOpenMultiplyVault'
 import { VaultAllowance, VaultAllowanceStatus } from 'components/vault/VaultAllowance'
 import { VaultChangesWithADelayCard } from 'components/vault/VaultChangesWithADelayCard'
 import { VaultErrors } from 'components/vault/VaultErrors'
@@ -12,6 +11,7 @@ import { VaultFormVaultTypeSwitch } from 'components/vault/VaultForm'
 import { VaultFormContainer } from 'components/vault/VaultFormContainer'
 import { VaultProxyContentBox, VaultProxyStatusCard } from 'components/vault/VaultProxy'
 import { VaultWarnings } from 'components/vault/VaultWarnings'
+import { SidebarOpenMultiplyVault } from 'features/multiply/open/sidebars/SidebarOpenMultiplyVault'
 import { extractGasDataFromState } from 'helpers/extractGasDataFromState'
 import { useFeatureToggle } from 'helpers/useFeatureToggle'
 import { useTranslation } from 'next-i18next'
