+++ features/multiply/open/sidebars/SidebarOpenMultiplyVaultOpenStage.tsx
